[refactor] 패키지 구조 변경; chess.Main -> chess.controller.ChessGame
board 와 peieces 패키지를 domain 패키지 안으로 넣음
미약하게나마 view, domain, controller 분리를 시도함
