+++ src/main/java/chess/domain/pieces/Pawn.java
@@ -1,4 +1,4 @@
-package chess.pieces;
+package chess.domain.pieces;
 
 public class Pawn extends Piece {
     public Pawn() {
