+++ src/test/java/chess/domain/pieces/PawnTest.java
@@ -1,4 +1,4 @@
-package chess.pieces;
+package chess.domain.pieces;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
