+++ src/test/java/chess/view/BoardViewTest.java
@@ -1,6 +1,6 @@
 package chess.view;
 
-import chess.board.Board;
+import chess.domain.board.Board;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
